requirements.txt
confluent-kafka==2.10.0
cryptography==45.0.3
elasticsearch==8.13.2
fastapi[standard]==0.115.4
ipython==8.29.0
pydantic==2.11.5
PyJWT==2.10.1
pytest-mock==3.14.1
pytest==8.3.5
strawberry-graphql[cli]==0.270.5
strawberry-graphql[fastapi]==0.270.5
